+++ src/index.js
@@ -2,22 +2,28 @@ import _ from "lodash";
 import fsp from "fs-promise";
 import rp from "request-promise";
 import cheerio from "cheerio";
+import iconv from "iconv-lite";
 
 const tabletojson = require('tabletojson');
 
 /**
  * Core Vertretungsplan class
  * @property {Array} table - Array containing substitution objects. The keys are determined by the table headings, but lowercased and singularized (Klasse(n) => klasse).
+ * @property {String} untis_ver - Version of Untis used
+ * @property {String} last_updated - Time, the plan was last updated
+ * @property {Array] messages - Array of MOTDs
  */
 class Vertretungsplan {
 
   /**
    *
    * @param url URL of the plan (like http://ohgspringe.de/phocadownload/plan/subst_001.htm), usually ends with subst_00x <br /> If prefixed with file:// the file will be loaded
+   * @param encoding Encoding of the Document. Defaults to ISO-8859-1
    */
-  constructor(url) {
+  constructor(url, encoding = "ISO-8859-1") {
     this.loaded = false;
     this.url = url;
+    this.encoding = encoding;
     if (url.indexOf('file://') == 0) { //test fallback
       this.file = url.substring(7);
     }
@@ -28,7 +34,13 @@ class Vertretungsplan {
    * @returns {Promise.<cheerio>} Promise
    */
   load() {
-    return this._fetchHtml().then((html) => {
+    const encoding = this.encoding;
+
+    return this._fetchHtml().then((body) => {
+        var decoded = iconv.decode(new Buffer(body), encoding);
+        return decoded;
+      })
+      .then((html) => {
         return cheerio.load(html);
       })
       .then((window) => {
@@ -42,12 +54,17 @@ class Vertretungsplan {
    * Loads HTML-String from file || url
    * @returns {Promise.<string>} Promise returning the html string
    * @private
-     */
+   */
   _fetchHtml() {
+
     if (this.file) {
-      return fsp.readFile(this.file, 'utf8');
+      return fsp.readFile(this.file);
     }
-    return rp.get(this.url);
+    return rp({
+      url: this.url,
+      method: "GET",
+      encoding: null
+    });
   }
 
   /**