+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Changelog
 
 - [v1.0.0](#v100)
+- [v1.0.1](#v101)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
@@ -14,3 +15,9 @@
  
  * minor fixes (README etc.)
  * Require node v4 as minimum
+
+### v1.0.2
+
+  * add encoding parameter in constructor (defaults to iso-8859-1)
+  * encode test file as iso-8859-1
+  * document meta informations